refactor: app bar query to functional component
add storybook for app bar query
