+++ Source/Plugins/Core/com.equella.core/js/__stories__/components/AppBarQuery.stories.tsx
@@ -1,0 +1,12 @@
+import * as React from "react";
+import { action } from "@storybook/addon-actions";
+import AppBarQuery from "../../tsrc/components/AppBarQuery";
+
+export default {
+  title: "AppBarQuery",
+  component: AppBarQuery
+};
+
+export const QueryText = () => (
+  <AppBarQuery query="" onChange={action("change")} />
+);
